+++ rentals/src/main/java/com/skyfleet/rentals/dto/PaymentRequestDTO.java
@@ -2,20 +2,8 @@ package com.skyfleet.rentals.dto;
 
 
 
-import java.time.LocalDateTime;
+import java.math.BigDecimal;
 
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.skyfleet.rentals.entity.Booking;
-import com.skyfleet.rentals.entity.PaymentStatus;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -26,7 +14,7 @@ public class PaymentRequestDTO extends BaseDTO {
 
     private Long bookingId;
 
-    private Double amountPaid;
+    private BigDecimal amountPaid;
 
 
      // Enum for payment status
