+++ rentals/src/main/java/com/skyfleet/rentals/dto/PenaltyResponseDTO.java
@@ -2,21 +2,13 @@ package com.skyfleet.rentals.dto;
 
 
 
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.skyfleet.rentals.entity.Booking;
 import com.skyfleet.rentals.entity.PenaltyReasonStatus;
 import com.skyfleet.rentals.entity.PenaltyStatus;
 
-import jakarta.persistence.Column;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
 import lombok.Getter;
 import lombok.Setter;
+
+import java.math.BigDecimal;
 
 
 @Getter
@@ -30,7 +22,7 @@ public class PenaltyResponseDTO extends BaseDTO {
 
     private PenaltyReasonStatus penaltyReason;
 
-    private Double penaltyAmount;
+    private BigDecimal penaltyAmount;
 
     private PenaltyStatus penaltyStatus; 
 }
