+++ rentals/src/main/java/com/skyfleet/rentals/dto/UndertakingResponseDTO.java
@@ -2,17 +2,10 @@ package com.skyfleet.rentals.dto;
 
 
 
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.skyfleet.rentals.entity.Booking;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
 import lombok.Getter;
 import lombok.Setter;
+
+import java.math.BigDecimal;
 
 
 @Getter
@@ -30,7 +23,7 @@ public class UndertakingResponseDTO extends BaseDTO {
     private Boolean isAccepted;
 
     
-    private Double depositAmount;
+    private BigDecimal depositAmount;
 
     
     private String damageClauseText;
